fix(desktop): split workspace panel lint owners
Problem:
PR #8748 exceeded the repository file-size and test-file-size budgets, so the lint workflow could not pass.

Root cause:
WorkspacePanel still owned tree-row and breadcrumb rendering, while one workspace DOM suite combined shared setup with an unrelated selection-isolation regression.

Fix:
Extract workspace tree rows and path breadcrumbs into focused components. Move the shared DOM setup into a test harness and register the selection-isolation regression as its own workspace suite.

Verification:
- go run ./tools/repolint
- pnpm test:workspace
- pnpm typecheck
- pnpm lint:hooks
- pnpm build
- git diff --check
