+++ desktop/frontend/src/__tests__/workspace-changes-errors.test.tsx
@@ -1,28 +1,9 @@
 // Run: tsx src/__tests__/workspace-changes-errors.test.tsx
 
-import { JSDOM } from "jsdom";
-import { registerHooks } from "node:module";
-import React from "react";
 import { act } from "react";
-import { createRoot } from "react-dom/client";
 import { workspaceFileIcon } from "../components/WorkspaceFileIcon";
-import { WorkspacePanel } from "../components/WorkspacePanel";
-import { LocaleProvider } from "../lib/i18n";
-import { resetWorkspaceTreeMemoryForTests } from "../lib/workspaceTreeMemory";
-import type { AppBindings } from "../lib/bridge";
-import type { DirEntry, FilePreview, GitCommitView, WireCompletionSummary, WorkspaceChangeDetailView, WorkspaceChangesView } from "../lib/types";
-
-// Markdown previews lazy-load MarkdownRenderer, whose KaTeX stylesheet belongs
-// to the same production chunk. Node's tsx loader has no CSS module support,
-// so map stylesheet imports to the existing empty asset stub in this DOM test.
-registerHooks({
-  resolve(specifier, context, nextResolve) {
-    if (specifier.endsWith(".css")) {
-      return nextResolve("./asset-stub-for-tests.ts", { ...context, parentURL: import.meta.url });
-    }
-    return nextResolve(specifier, context);
-  },
-});
+import type { DirEntry, FilePreview, WorkspaceChangeDetailView, WorkspaceChangesView } from "../lib/types";
+import { flushPromises, renderFilesWorkspace, renderWorkspace, waitFor } from "./workspace-panel-test-harness";
 
 let passed = 0;
 let failed = 0;
@@ -37,243 +18,7 @@ function ok(value: boolean, label: string) {
   }
 }
 
-function flushPromises(): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, 0));
-}
-
-async function waitFor(label: string, predicate: () => boolean) {
-  for (let attempt = 0; attempt < 20; attempt += 1) {
-    await act(async () => {
-      await flushPromises();
-    });
-    if (predicate()) return;
-  }
-  throw new Error(`timed out waiting for ${label}`);
-}
-
-class TestResizeObserver {
-  observe() {}
-  unobserve() {}
-  disconnect() {}
-}
-
-function installDom() {
-  const dom = new JSDOM("<!doctype html><html><body><div id=\"root\"></div></body></html>", {
-    pretendToBeVisual: true,
-    url: "http://localhost/",
-  });
-  (globalThis as typeof globalThis & { IS_REACT_ACT_ENVIRONMENT: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
-  globalThis.window = dom.window as unknown as Window & typeof globalThis;
-  globalThis.document = dom.window.document;
-  Object.defineProperty(globalThis, "navigator", { configurable: true, value: dom.window.navigator });
-  globalThis.Node = dom.window.Node;
-  globalThis.Element = dom.window.Element;
-  globalThis.HTMLElement = dom.window.HTMLElement;
-  globalThis.Event = dom.window.Event;
-  globalThis.CustomEvent = dom.window.CustomEvent;
-  globalThis.KeyboardEvent = dom.window.KeyboardEvent;
-  globalThis.MouseEvent = dom.window.MouseEvent;
-  globalThis.PointerEvent = dom.window.MouseEvent as unknown as typeof PointerEvent;
-  globalThis.MutationObserver = dom.window.MutationObserver;
-  globalThis.ResizeObserver = TestResizeObserver;
-  dom.window.ResizeObserver = TestResizeObserver;
-  globalThis.localStorage = dom.window.localStorage;
-  globalThis.requestAnimationFrame = dom.window.requestAnimationFrame.bind(dom.window);
-  globalThis.cancelAnimationFrame = dom.window.cancelAnimationFrame.bind(dom.window);
-  (dom.window.HTMLElement.prototype as unknown as { attachEvent: () => void }).attachEvent = () => {};
-  (dom.window.HTMLElement.prototype as unknown as { detachEvent: () => void }).detachEvent = () => {};
-  Object.defineProperty(dom.window.HTMLElement.prototype, "scrollIntoView", { configurable: true, value: () => {} });
-  Object.defineProperty(dom.window.HTMLElement.prototype, "offsetWidth", {
-    configurable: true,
-    get: () => 320,
-  });
-  Object.defineProperty(dom.window.HTMLElement.prototype, "offsetHeight", {
-    configurable: true,
-    get: function offsetHeight(this: HTMLElement) {
-      return this.classList.contains("workspace-tree") ? 300 : this.dataset.index ? 24 : 0;
-    },
-  });
-  Object.defineProperty(dom.window.HTMLElement.prototype, "getBoundingClientRect", {
-    configurable: true,
-    value: function getBoundingClientRect(this: HTMLElement) {
-      const width = 320;
-      const height = this.classList.contains("workspace-tree") ? 300 : this.dataset.index ? 24 : 0;
-      return {
-        x: 0,
-        y: 0,
-        top: 0,
-        left: 0,
-        right: width,
-        bottom: height,
-        width,
-        height,
-        toJSON: () => ({}),
-      } as DOMRect;
-    },
-  });
-  return dom;
-}
-
-async function renderWorkspace(
-  changes: WorkspaceChangesView,
-  options: { creationMode?: boolean; history?: GitCommitView[]; detail?: WorkspaceChangeDetailView; completionSummary?: WireCompletionSummary } = {},
-) {
-  resetWorkspaceTreeMemoryForTests();
-  const dom = installDom();
-  window.go = {
-    main: {
-      App: {
-        ListDirForTab: async () => [],
-        WorkspaceGitHistory: async () => options.history ?? [],
-        WorkspaceChanges: async () => changes,
-        WorkspaceChangeDetail: async () => options.detail ?? {},
-        ReadFileForTab: async (_tabID, path) => ({ path, body: "", size: 0, truncated: false, binary: false }),
-      } as Partial<AppBindings> as AppBindings,
-    },
-  };
-  const rootEl = document.getElementById("root");
-  if (!rootEl) throw new Error("missing root");
-  const root = createRoot(rootEl);
-  await act(async () => {
-    root.render(
-      <LocaleProvider>
-        <WorkspacePanel
-          open
-          tabId="tab-a"
-          cwd="/repo"
-          maximized={false}
-          initialViewMode="changed"
-          creationMode={options.creationMode}
-          completionSummary={options.completionSummary}
-          onClose={() => {}}
-          onToggleMaximized={() => {}}
-        />
-      </LocaleProvider>,
-    );
-    await flushPromises();
-  });
-  await waitFor("workspace changes", () => Boolean(document.querySelector(".workspace-preview__body")));
-  return { dom, root };
-}
-
-async function renderFilesWorkspace(methods: Partial<AppBindings>, props: Partial<Parameters<typeof WorkspacePanel>[0]> = {}) {
-  resetWorkspaceTreeMemoryForTests();
-  const dom = installDom();
-  window.go = {
-    main: {
-      App: {
-        ListDirForTab: async () => [],
-        SearchFileRefsForTab: async () => [],
-        WorkspaceGitHistory: async () => [],
-        WorkspaceChanges: async () => ({ files: [], gitAvailable: true }),
-        WorkspaceChangeDetail: async () => ({}),
-        ReadFileForTab: async (_tabID, path) => ({ path, body: "", size: 0, truncated: false, binary: false }),
-        ...methods,
-      } as Partial<AppBindings> as AppBindings,
-    },
-  };
-  const rootEl = document.getElementById("root");
-  if (!rootEl) throw new Error("missing root");
-  const root = createRoot(rootEl);
-  let currentProps: Parameters<typeof WorkspacePanel>[0] = {
-    open: true,
-    tabId: "tab-a",
-    cwd: "/repo",
-    maximized: false,
-    initialViewMode: "files",
-    onClose: () => {},
-    onToggleMaximized: () => {},
-    ...props,
-  };
-  const rerender = async (nextProps: Partial<Parameters<typeof WorkspacePanel>[0]> = {}) => {
-    currentProps = { ...currentProps, ...nextProps };
-    await act(async () => {
-      root.render(
-        <LocaleProvider>
-          <WorkspacePanel {...currentProps} />
-        </LocaleProvider>,
-      );
-      await flushPromises();
-    });
-  };
-  await rerender();
-  return { dom, root, rerender };
-}
-
 console.log("\nworkspace changes git errors");
-
-{
-  const changeDetailPaths: string[] = [];
-  const { dom, root, rerender } = await renderFilesWorkspace({
-    ListDirForTab: async (_tabId, dir) => dir === "" ? [{ name: "notes.txt", isDir: false }] : [],
-    WorkspaceChanges: async () => ({
-      files: [{ path: "src/changed.ts", sources: ["git"], gitStatus: "M" }],
-      gitAvailable: true,
-    }),
-    WorkspaceChangeDetail: async (_tabId, path) => {
-      changeDetailPaths.push(path);
-      return { source: "git", diff: "changed diff" };
-    },
-    ReadFileForTab: async (_tabId, path) => ({ path, body: "file preview", size: 12, truncated: false, binary: false }),
-  });
-
-  await waitFor("files view entry", () => document.body.textContent?.includes("notes.txt") === true);
-  await act(async () => {
-    document.querySelector<HTMLButtonElement>('[data-workspace-path="notes.txt"]')
-      ?.dispatchEvent(new window.MouseEvent("click", { bubbles: true }));
-    await flushPromises();
-  });
-  await waitFor("selected file preview", () => document.body.textContent?.includes("file preview") === true);
-
-  await rerender({ initialViewMode: "changed" });
-  await waitFor("changes overview", () => document.body.textContent?.includes("changed.ts") === true);
-
-  ok(changeDetailPaths.length === 0, "entering Changes does not request diff detail for the Files selection");
-  ok(document.querySelector(".workspace-panel--changed-overview") !== null, "entering Changes opens its overview instead of leaking the Files selection");
-
-  await act(async () => {
-    document.querySelector<HTMLButtonElement>(".workspace-change")
-      ?.dispatchEvent(new window.MouseEvent("click", { bubbles: true }));
-    await flushPromises();
-  });
-  await waitFor("selected change detail", () => changeDetailPaths.includes("src/changed.ts"));
-
-  await rerender({ initialViewMode: "files" });
-  await waitFor("restored file preview", () => document.body.textContent?.includes("file preview") === true);
-
-  ok(changeDetailPaths.every((path) => path === "src/changed.ts"), "Changes requests only its own selected path");
-  ok(document.body.textContent?.includes("notes.txt") === true, "returning to Files restores the prior file selection after viewing a change");
-
-  await rerender({
-    initialViewMode: "changed",
-    changeListRequest: {
-      id: 1,
-      changes: [{ key: "scoped-change", path: "src/changed.ts", meta: "M", time: "", detail: "changed" }],
-    },
-  });
-  await waitFor("scoped Changes list", () => document.body.textContent?.includes("Session changes") === true);
-  await rerender({ initialViewMode: "files", changeListRequest: null });
-  await waitFor("file preview after scoped Changes", () => document.body.textContent?.includes("file preview") === true);
-  ok(document.body.textContent?.includes("notes.txt") === true, "scoped Changes requests preserve the Files preview state");
-
-  await rerender({ initialViewMode: "changed" });
-  await waitFor("change before scope switch", () => document.body.textContent?.includes("changed.ts") === true);
-  await act(async () => {
-    document.querySelector<HTMLButtonElement>(".workspace-change")
-      ?.dispatchEvent(new window.MouseEvent("click", { bubbles: true }));
-    await flushPromises();
-  });
-  await waitFor("change selected before scope switch", () => changeDetailPaths.includes("src/changed.ts"));
-  const callsBeforeScopeSwitch = changeDetailPaths.length;
-  await rerender({ workspaceScopeKey: "other-session" });
-  await waitFor("new scope Changes overview", () => document.querySelector(".workspace-panel--changed-overview") !== null);
-  ok(changeDetailPaths.length === callsBeforeScopeSwitch, "workspace scope changes do not request the previous scope's selected change");
-
-  await act(async () => {
-    root.unmount();
-  });
-  dom.window.close();
-}
 
 {
   const { dom, root, rerender } = await renderFilesWorkspace({}, { open: false });
